moved main server files to root
