+++ index.js
@@ -2,7 +2,7 @@
 // login heroku `heroku login`
 //From the top level of repo use command `git subtree push --prefix server heroku-backend master`
 
-const server = require('./api/server.js');
+const server = require('./server/api/server.js');
 const port = process.env.PORT || 3000;
 
 // start the server
